+++ src/main/java/controller/menu/Menu.java
@@ -1,6 +1,6 @@
 package controller.menu;
 
-import model.ChessGame;
+import model.game.ChessGame;
 
 public interface Menu {
