+++ src/main/java/database/ChessGameDao.java
@@ -4,7 +4,7 @@ import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import model.Camp;
+import model.game.Camp;
 
 public class ChessGameDao {
