+++ src/main/java/dto/ChessGameDto.java
@@ -1,7 +1,7 @@
 package dto;
 
 import java.util.Map;
-import model.Camp;
+import model.game.Camp;
 import model.piece.Piece;
 import model.position.Position;
