+++ src/main/java/model/game/ChessBoard.java
@@ -1,4 +1,4 @@
-package model;
+package model.game;
 
 import java.util.Collections;
 import java.util.EnumMap;
