+++ src/main/java/model/game/ChessStatus.java
@@ -1,4 +1,4 @@
-package model;
+package model.game;
 
 public enum ChessStatus {
     INIT,
