+++ src/main/java/model/game/Result.java
@@ -1,4 +1,4 @@
-package model;
+package model.game;
 
 public enum Result {
