+++ src/main/java/model/piece/Piece.java
@@ -2,7 +2,7 @@ package model.piece;
 
 import java.util.Objects;
 import java.util.Set;
-import model.Camp;
+import model.game.Camp;
 import model.position.Moving;
 import model.position.Position;
