+++ src/test/java/model/ChessGameTest.java
@@ -3,6 +3,7 @@ package model;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import model.game.ChessGame;
 import model.position.Moving;
 import model.position.Position;
 import org.junit.jupiter.api.DisplayName;
