+++ src/test/java/model/piece/KingTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.stream.Stream;
-import model.Camp;
+import model.game.Camp;
 import model.position.Moving;
 import model.position.Position;
 import org.junit.jupiter.api.DisplayName;
